fix: filter exposed feature flags
